+++ fe-house-p/src/app/HomeMovilNavbar/page.js
@@ -1,10 +1,10 @@
 import React from "react";
-import { MovilNavbar } from "@/components/MovilNavBar";
+
 
 export default function HomeMovilNavbar() {
     return (
         <div>
-            <MovilNavbar />
+            
         </div>
     );
 }
