+++ fe-house-p/src/app/SignUp/page.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+import { SignUpComponent } from "@/components/SignUpComponent";
+export default function SignUp() {
+    return (
+      <>
+        <SignUpComponent />
+      </>
+    );
+}
