+++ fe-house-p/src/assets/images/index.jsx
@@ -1,0 +1,4 @@
+import meme from './meme.png'
+import en_construccion from './en_construccion.jpg'
+
+export { meme, en_construccion};
