--- fe-house-p/src/components/ExampleComponent/index.js
@@ -1,3 +1,0 @@
-import ExampleComponent from "./ExampleComponent";
-
-export { ExampleComponent };
