--- fe-house-p/src/components/ExampleMuiUi/index.js
@@ -1,3 +1,0 @@
-import ExampleMuiUi from "./ExampleMuiUi";
-
-export { ExampleMuiUi };
