+++ fe-house-p/src/components/LoginComponent/index.jsx
@@ -1,0 +1,3 @@
+import LoginComponent from './LoginComponent';
+
+export { LoginComponent };
