+++ fe-house-p/src/components/SignUpComponent/index.jsx
@@ -1,0 +1,3 @@
+import SignUpComponent from './SignUpComponent';
+
+export { SignUpComponent };
